add metadata and render modules
